Added AuthServiceImpl LoginResponse , Logout

+++ microservices/auth-service/src/main/java/dev/bengi/authservice/model/RefreshToken.java
@@ -16,6 +16,6 @@ public record RefreshToken(
         String token,
         Instant issuedAt,
         String usage,
-        Long resourceId,
+        Integer resourceId,
         boolean isExpired) {
 }
